revert tests

+++ test/unit/furnace/FurnaceHealthUnitCalcsFuzz.t.sol
@@ -6,8 +6,8 @@ import {Base_Test} from "../../Base.t.sol";
 import "src/contracts/libraries/BasicMath.sol";
 
 contract FurnaceHealthUnitCalcsTestFuzz is Base_Test {
-    using BasicMath for int;
-    using BasicMath for uint;
+    using BasicMath for int256;
+    using BasicMath for uint256;
 
     function setUp() public virtual override {
         Base_Test.setUp();
@@ -24,23 +24,18 @@ contract FurnaceHealthUnitCalcsTestFuzz is Base_Test {
     ) public view {
         vm.assume(quantity < 10000000000000000e18);
         vm.assume(spotPrice < 10000000000000000e18);
-        vm.assume(
-            uint256(quantity).mul(uint256(perpEntryPrice)) <
-                10000000000000000e18
-        );
+        vm.assume(uint256(quantity).mul(uint256(perpEntryPrice)) < 10000000000000000e18);
         vm.assume(perpPrice < 10000000000000000e18);
         vm.assume(perpEntryPrice < 10000000000000000e18);
         vm.assume(
-            int256(initCumFunding) < 10000000000000000e18 &&
-                int256(initCumFunding) > -10000000000000000e18
+            int256(initCumFunding) < 10000000000000000e18
+                && int256(initCumFunding) > -10000000000000000e18
         );
         vm.assume(
-            int256(currentCumFunding) < 10000000000000000e18 &&
-                int256(currentCumFunding) > -10000000000000000e18
+            int256(currentCumFunding) < 10000000000000000e18
+                && int256(currentCumFunding) > -10000000000000000e18
         );
-        vm.assume(
-            abs(int256(initCumFunding) - int256(currentCumFunding)) < 100000e18
-        );
+        vm.assume(abs(int256(initCumFunding) - int256(currentCumFunding)) < 100000e18);
 
         mockFurnaceUnitTest.calculateSpreadHealth(
             defaults.wbtcSpreadPenalty().initial, // initial
@@ -64,23 +59,18 @@ contract FurnaceHealthUnitCalcsTestFuzz is Base_Test {
         vm.assume(quantity < 10000000000000000e18);
         vm.assume(quantity < 10000000000000000e18);
         vm.assume(spotPrice < 10000000000000000e18);
-        vm.assume(
-            uint256(quantity).mul(uint256(perpEntryPrice)) <
-                10000000000000000e18
-        );
+        vm.assume(uint256(quantity).mul(uint256(perpEntryPrice)) < 10000000000000000e18);
         vm.assume(perpPrice < 10000000000000000e18);
         vm.assume(perpEntryPrice < 10000000000000000e18);
         vm.assume(
-            int256(initCumFunding) < 10000000000000000e18 &&
-                int256(initCumFunding) > -10000000000000000e18
+            int256(initCumFunding) < 10000000000000000e18
+                && int256(initCumFunding) > -10000000000000000e18
         );
         vm.assume(
-            int256(currentCumFunding) < 10000000000000000e18 &&
-                int256(currentCumFunding) > -10000000000000000e18
+            int256(currentCumFunding) < 10000000000000000e18
+                && int256(currentCumFunding) > -10000000000000000e18
         );
-        vm.assume(
-            abs(int256(initCumFunding) - int256(currentCumFunding)) < 100000e18
-        );
+        vm.assume(abs(int256(initCumFunding) - int256(currentCumFunding)) < 100000e18);
 
         mockFurnaceUnitTest.calculateSpreadHealth(
             defaults.wethSpreadPenalty().initial, // initial
@@ -104,23 +94,18 @@ contract FurnaceHealthUnitCalcsTestFuzz is Base_Test {
         vm.assume(quantity < 10000000000000000e18);
         vm.assume(quantity < 10000000000000000e18);
         vm.assume(spotPrice < 10000000000000000e18);
-        vm.assume(
-            uint256(quantity).mul(uint256(perpEntryPrice)) <
-                10000000000000000e18
-        );
+        vm.assume(uint256(quantity).mul(uint256(perpEntryPrice)) < 10000000000000000e18);
         vm.assume(perpPrice < 10000000000000000e18);
         vm.assume(perpEntryPrice < 10000000000000000e18);
         vm.assume(
-            int256(initCumFunding) < 10000000000000000e18 &&
-                int256(initCumFunding) > -10000000000000000e18
+            int256(initCumFunding) < 10000000000000000e18
+                && int256(initCumFunding) > -10000000000000000e18
         );
         vm.assume(
-            int256(currentCumFunding) < 10000000000000000e18 &&
-                int256(currentCumFunding) > -10000000000000000e18
+            int256(currentCumFunding) < 10000000000000000e18
+                && int256(currentCumFunding) > -10000000000000000e18
         );
-        vm.assume(
-            abs(int256(initCumFunding) - int256(currentCumFunding)) < 100000e18
-        );
+        vm.assume(abs(int256(initCumFunding) - int256(currentCumFunding)) < 100000e18);
 
         mockFurnaceUnitTest.calculateSpreadHealth(
             defaults.wbtcSpreadPenalty().maintenance, // maintenance
@@ -144,23 +129,18 @@ contract FurnaceHealthUnitCalcsTestFuzz is Base_Test {
         vm.assume(quantity < 10000000000000000e18);
         vm.assume(quantity < 10000000000000000e18);
         vm.assume(spotPrice < 10000000000000000e18);
-        vm.assume(
-            uint256(quantity).mul(uint256(perpEntryPrice)) <
-                10000000000000000e18
-        );
+        vm.assume(uint256(quantity).mul(uint256(perpEntryPrice)) < 10000000000000000e18);
         vm.assume(perpPrice < 10000000000000000e18);
         vm.assume(perpEntryPrice < 10000000000000000e18);
         vm.assume(
-            int256(initCumFunding) < 10000000000000000e18 &&
-                int256(initCumFunding) > -10000000000000000e18
+            int256(initCumFunding) < 10000000000000000e18
+                && int256(initCumFunding) > -10000000000000000e18
         );
         vm.assume(
-            int256(currentCumFunding) < 10000000000000000e18 &&
-                int256(currentCumFunding) > -10000000000000000e18
+            int256(currentCumFunding) < 10000000000000000e18
+                && int256(currentCumFunding) > -10000000000000000e18
         );
-        vm.assume(
-            abs(int256(initCumFunding) - int256(currentCumFunding)) < 100000e18
-        );
+        vm.assume(abs(int256(initCumFunding) - int256(currentCumFunding)) < 100000e18);
 
         mockFurnaceUnitTest.calculateSpreadHealth(
             defaults.wethSpreadPenalty().maintenance, // maintenance
